requirements.txt
Pillow==9.5.0
PyPDF2==3.0.1
pytube==12.1.3
reportlab==3.6.12
